+++ Changes.md
@@ -24,6 +24,7 @@ Fixes
 
 - Expression, OSLCode : Fixed line numbers reported in OSL parse errors.
 - PathColumn : Fixed display of swatches for cells containing `Color4fData`.
+- Arnold : Fixed "Flush Cache" menu items to work with renders being performed by an InteractiveRender node (rather than an InteractiveArnoldRender node).
 
 API
 ---
@@ -35,6 +36,7 @@ API
 - MultiLineTextWidget, CodeWidget : Added the ability to show line numbers by passing `lineNumbersVisible = True` to the constructor.
 - MultiLineTextWidget : Added `setLineNumbersVisible()` and `getLineNumbersVisible()`
 - PathListingWidget : Added `expandTo()` method.
+- InteractiveRender : Added `command()` method to run a renderer command.
 
 Breaking Changes
 ----------------